requirements.txt
fastapi==0.112.1
celery== 5.4.0
uvicorn == 0.27.0
redis == 5.0.8
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.6
python-dotenv == 1.0.1